Made all REPLs accept multiline input before evaluating a form
The behaviour of other REPLs should match the worker REPL now. This
also fixes a problem where the worker REPL would go into an infinite
loop if you started a line with ). It also fixes an issue where the
web REPL allowed imbalanced parentheses at the start of an expression.

This fixes #277